Move cache-related code to a separate module
